Fixed the SiteNav to show up on all the rendered pages => However, two SiteNav bars show up on home page
