+++ client/src/components/Home/style.css
@@ -1,4 +1,11 @@
 #cardWrapper {
-    width: 600px;
-    margin: auto;
+    max-width: 500px;
+    margin: 30px auto;
+    padding: 20px;
+    border: 0.5px;
+    border-style: solid;
+    border-radius: 10px;
+    border-color: grey;
+    background-color: #fff;
+    box-shadow: 10px 10px 5px grey;
 }
