+++ client/src/components/pages/Game/Game.js
@@ -1,0 +1,18 @@
+// ***************** Game Selection Page *************** //
+// This page is basic lore of the game.  This will provide overall reason for the player to do what 
+// they will be doing and why it is important for certain actions to be done. 
+// Rules / Gameplay mechanics will be explained on this page
+// This page is accessible from any of the pages => not tied to any specifc flow
+
+import React from "react";
+
+const Game = () => (
+    <div>
+      <h1>REMNANT: The Lore</h1>
+      <p>
+        What is going on in solar system that is making your people choose you to be save the day?
+      </p>
+    </div>
+  );
+  
+  export default Game;
